requirements.txt
# Browser automation
playwright>=1.40.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0
Pillow>=10.0.0
numpy>=1.24.0

# Telegram bot
python-telegram-bot>=21.0

# LLM APIs
anthropic>=0.18.0
requests>=2.31.0

# Config
python-dotenv>=1.0.0
PyYAML>=6.0
